fix: update indexer package
