+++ packages/contracts/package.json
@@ -34,7 +34,6 @@
     "dotenv": "^10.0.0",
     "ethereum-waffle": "^3.4.0",
     "ethers": "^5.7.2",
-    "hardhat": "^2.22.1",
     "hardhat-gas-reporter": "^1.0.8",
     "ts-node": "^10.4.0",
     "typescript": "^4.5.2",
