+++ packages/indexer/src/api/endpoints/execute/get-execute-list/v5.ts
@@ -807,7 +807,7 @@ export const getExecuteListV5Options: RouteOptions = {
               }
 
               case "seaport-v1.6": {
-                if (!["reservoir", "opensea", "looks-rare"].includes(params.orderbook)) {
+                if (!["reservoir", "opensea"].includes(params.orderbook)) {
                   return errors.push({ message: "Unsupported orderbook", orderIndex: i });
                 }
