+++ packages/sdk/src/seaport-v1.6/addresses.ts
@@ -2,4 +2,15 @@ import { ChainIdToAddress, Network } from "../utils";
 
 export const Exchange: ChainIdToAddress = {
   [Network.Ethereum]: "0x0000000000000068f116a894984e2db1123eb395",
+  [Network.EthereumSepolia]: "0x0000000000000068f116a894984e2db1123eb395",
+  [Network.Polygon]: "0x0000000000000068f116a894984e2db1123eb395",
+  [Network.Mumbai]: "0x0000000000000068f116a894984e2db1123eb395",
+  [Network.Optimism]: "0x0000000000000068f116a894984e2db1123eb395",
+  [Network.Arbitrum]: "0x0000000000000068f116a894984e2db1123eb395",
+  [Network.ArbitrumNova]: "0x0000000000000068f116a894984e2db1123eb395",
+  [Network.Base]: "0x0000000000000068f116a894984e2db1123eb395",
+  [Network.BaseSepolia]: "0x0000000000000068f116a894984e2db1123eb395",
+  [Network.Avalanche]: "0x0000000000000068f116a894984e2db1123eb395",
+  [Network.Bsc]: "0x0000000000000068f116a894984e2db1123eb395",
+  [Network.Zora]: "0x0000000000000068f116a894984e2db1123eb395",
 };
